+++ aiogram_nats/core/interactors/mailing/start_mailing.py
@@ -1,13 +1,12 @@
 from aiogram_nats.core.entities.mailing import Mailing, ScheduledMailing
-from aiogram_nats.core.interfaces.interfaces.mailing import MailingStarter
-from aiogram_nats.core.interfaces.interfaces.scheduler import Scheduler
+from aiogram_nats.core.interfaces.interfaces.scheduler import Scheduler, Task
 
 
 class StartMailing:
 
     """Class is responsible for starting a mailing."""
 
-    def __init__(self, mailing_starter: MailingStarter) -> None:
+    def __init__(self, mailing_starter: Task) -> None:
         self.mailing_starter = mailing_starter
 
     async def __call__(self, mailing: Mailing) -> None:
@@ -20,14 +19,14 @@ class StartMailing:
         Returns :
             None: This function does not return anything.
         """
-        await self.mailing_starter.start(mailing)
+        await self.mailing_starter(mailing)
 
 
 class ScheduleMailing:
 
     """Class is responsible for scheduling a mailing."""
 
-    def __init__(self, mailing_starter: MailingStarter, scheduler: Scheduler) -> None:
+    def __init__(self, mailing_starter: Task, scheduler: Scheduler) -> None:
         self.mailing_starter = mailing_starter
         self.scheduler = scheduler
 
@@ -42,5 +41,4 @@ class ScheduleMailing:
             str: The scheduling ID of the mailing.
 
         """
-        starter = self.mailing_starter.start
-        return await self.scheduler.schedule(mailing, starter)
+        return await self.scheduler.schedule(mailing, self.mailing_starter)
