+++ aiogram_nats/core/interactors/messages/__init__.py
@@ -1,5 +1,4 @@
-from .deletion_scheduled import ScheduleMessageDeletion
-from .send_scheduled import ScheduleMessageSend
+from .scheduled import ScheduleMessageDeletion, ScheduleMessageSend
 
 __all__ = [
     "ScheduleMessageSend",
